# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TaskManager DOC API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user/tasks": {
            "get": {
                "tags": [
                    "Пользовательские функции (Данные Заданий)"
                ],
                "summary": "Список Заданий авторизированного пользователя",
                "operationId": "d4d32330e747d424c3635dea25a9ebde",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Название Задания",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "content",
                        "in": "query",
                        "description": "Описание Задания",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Статус задания,\n *          1-Создано\n *          2-Исполнено\n *          3-Удалено",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "created_from",
                        "in": "query",
                        "description": "Дата создания с",
                        "required": false,
                        "schema": {
                            "type": "date-time"
                        }
                    },
                    {
                        "name": "created_to",
                        "in": "query",
                        "description": "Дата создания до",
                        "required": false,
                        "schema": {
                            "type": "date-time"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Переустановить что-то"
                                                    },
                                                    "content": {
                                                        "type": "text",
                                                        "example": "Способ переустановки"
                                                    },
                                                    "status": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2010-10-22"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Пользовательские функции (Данные Заданий)"
                ],
                "summary": "Создание 'Задания', автоматические прокидывает user_id и status=1",
                "operationId": "70d6e85d9f3d9527401ce4f50b6dc6a9",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "title": {
                                                "type": "string",
                                                "example": "Переустановить MacOs"
                                            },
                                            "content": {
                                                "type": "text",
                                                "example": "Действия по заданию"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Переустановить MacOs"
                                        },
                                        "content": {
                                            "type": "text",
                                            "example": "Действия по заданию"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2010-10-22"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unautoruzed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/tasks/{task}": {
            "get": {
                "tags": [
                    "Пользовательские функции (Данные Заданий)"
                ],
                "summary": "Просмотр Задания авторизированного пользователя",
                "operationId": "e73bba40382b16f61edeef541ad80deb",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "description": "id task",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Переустановить MacOs"
                                        },
                                        "content": {
                                            "type": "text",
                                            "example": "Действия по заданию"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2010-10-22"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Пользовательские функции (Данные Заданий)"
                ],
                "summary": "Удаление 'Задание', принадлажащего авторизированному пользователю",
                "operationId": "07009363d214107156b7d4d8ce572eb9",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "description": "id task",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": " Task Deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Пользовательские функции (Данные Заданий)"
                ],
                "summary": "Обновление данных Задания принадлажащего авторизированному пользователю",
                "operationId": "fd4c353311e79baac6fe7a3984ba3a72",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "description": "id task",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "title": {
                                                "type": "string",
                                                "example": "Переустановить MacOs"
                                            },
                                            "content": {
                                                "type": "text",
                                                "example": "Действия по заданию"
                                            },
                                            "status": {
                                                "type": "integer",
                                                "example": 1
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Переустановить MacOs"
                                                    },
                                                    "content": {
                                                        "type": "text",
                                                        "example": "Действия по заданию"
                                                    },
                                                    "status": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2010-10-22"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unautoruzed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}